% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{classifier_fold1_small}
\alias{classifier_fold1_small}
\title{small classifier for training fold 1}
\format{An object of class \code{ranger} of length 12.}
\usage{
classifier_fold1_small
}
\description{
small classifier for training fold 1
}
\keyword{datasets}

